requirements.txt
flask==3.0.2
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
python-dotenv==1.0.0
argon2-cffi==23.1.0
flask-talisman==1.0.0
email_validator>=1.3.0
wtforms
pip install -r requirements.txt
